hwclock: add --systz option to set system clock from itself
Since the system clock time is already set from the hardware clock by
the kernel, there's no particular need to read the hardware clock
again.

This option sets the system clock using itself as a reference if the
hardware clock was in local time.  The resulting system clock time
is in UTC, with the kernel timezone set to the difference.