hwclock: don't call hwclock.sh on startup, drop hwclockfirst.sh
Since we set the system clock from a udev rule, we don't need to do it
from an init script and we certainly don't need to do it twice!

Remove the start links for the hwclock.sh script (noting in a comment
that it is usual that they are missing), and remove the hwclockfirst.sh
script entirely.

Make sure it's sorted out on upgrade.