Revert "Avoid starting uuidd when having socket activation"
This reverts commit 0005ffa80dbe299c9cc3cfaefd05827ebff8c80e.

This is already too hacky and we need an additional hack for
this to work. The uuidd.socket needs to be started and
dh-systemd (or anything else) doesn't have any way
to only start the socket, without starting the service.
Reverting and taking the penalty seems better then to keep
adding more hacks to get this working.