Override localstatedir to /run instead of /var
Debhelper sets localstatedir to /var. Upstream util-linux defaults to
using /run for this.

The only thing it's used for is the uuidd state files which the package
previously put under /var/run.

Set it to /run for now so we get the benefit of automatic cleanup (as
well as not deviating from upstream).