Revert "Trigger update of initramfs on upgrades"
This reverts commit 7fe640d674d35b19dae4a19cdb28c8e70c0b0867.

The new initramfs-tools did not make it into Jessie.
This triggers a bug in live-tools, which dropping the update-initramfs
is one way of working around.
For the initramfs-tools version (0.117 & 0.118) stuck in unstable we
can hopefully make it not have to pull in parts of util-linux
during the Stretch development cycle (and even if not, an outdated
initramfs isn't the end of the world).

Closes: #773354